Changing the data type from double to int because the code gets errors and it wont run

+++ Prog6221_Part1_st10180168_GabrielMoraka/Functions.cs
@@ -12,7 +12,7 @@ namespace Prog6221_Part1_st10180168_GabrielMoraka
     public    class MethodsAndFunctions
         {
             public string[] ingredients;
-            public double[] quantities;
+            public int[] quantities;
             public string[] units;
             public string[] steps;
